+++ lib/src/controls/flutter_deck_controls_notifier.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import 'package:flutter/widgets.dart';
 import 'package:flutter_deck/src/controls/actions/actions.dart';
+import 'package:flutter_deck/src/controls/autoplay/autoplay.dart';
 import 'package:flutter_deck/src/controls/fullscreen/fullscreen.dart';
 import 'package:flutter_deck/src/flutter_deck_router.dart';
 import 'package:flutter_deck/src/widgets/internal/drawer/drawer.dart';
@@ -14,15 +15,18 @@ class FlutterDeckControlsNotifier
     implements FlutterDeckFullscreenManagerBase {
   /// Creates a [FlutterDeckControlsNotifier].
   FlutterDeckControlsNotifier({
+    required FlutterDeckAutoplayNotifier autoplayNotifier,
     required FlutterDeckDrawerNotifier drawerNotifier,
     required FlutterDeckMarkerNotifier markerNotifier,
     required FlutterDeckFullscreenManagerBase fullscreenManager,
     required FlutterDeckRouter router,
-  })  : _drawerNotifier = drawerNotifier,
+  })  : _autoplayNotifier = autoplayNotifier,
+        _drawerNotifier = drawerNotifier,
         _markerNotifier = markerNotifier,
         _fullscreenManager = fullscreenManager,
         _router = router;
 
+  final FlutterDeckAutoplayNotifier _autoplayNotifier;
   final FlutterDeckDrawerNotifier _drawerNotifier;
   final FlutterDeckMarkerNotifier _markerNotifier;
   final FlutterDeckFullscreenManagerBase _fullscreenManager;
@@ -38,18 +42,21 @@ class FlutterDeckControlsNotifier
 
   /// Go to the next slide.
   void next() {
+    _autoplayNotifier.pause();
     _router.next();
     notifyListeners();
   }
 
   /// Go to the previous slide.
   void previous() {
+    _autoplayNotifier.pause();
     _router.previous();
     notifyListeners();
   }
 
   /// Toggle the navigation drawer.
   void toggleDrawer() {
+    _autoplayNotifier.pause();
     _drawerNotifier.toggle();
     notifyListeners();
   }
@@ -61,6 +68,7 @@ class FlutterDeckControlsNotifier
   /// * [GoPreviousIntent]
   /// * [ToggleDrawerIntent]
   void toggleMarker() {
+    _autoplayNotifier.pause();
     _markerNotifier.toggle();
 
     _disabledIntents = {