+++ lib/src/flutter_deck_app.dart
@@ -105,6 +105,7 @@ class _FlutterDeckAppState extends State<FlutterDeckApp> {
   late FlutterDeckRouter _flutterDeckRouter;
   late GoRouter _router;
 
+  late FlutterDeckAutoplayNotifier _autoplayNotifier;
   late FlutterDeckControlsNotifier _controlsNotifier;
   late FlutterDeckDrawerNotifier _drawerNotifier;
   late FlutterDeckMarkerNotifier _markerNotifier;
@@ -116,7 +117,9 @@ class _FlutterDeckAppState extends State<FlutterDeckApp> {
 
     _buildRouter();
 
+    _autoplayNotifier = FlutterDeckAutoplayNotifier(router: _flutterDeckRouter);
     _controlsNotifier = FlutterDeckControlsNotifier(
+      autoplayNotifier: _autoplayNotifier,
       drawerNotifier: _drawerNotifier = FlutterDeckDrawerNotifier(),
       markerNotifier: _markerNotifier = FlutterDeckMarkerNotifier(),
       fullscreenManager: FlutterDeckFullscreenManager(),
@@ -157,6 +160,7 @@ class _FlutterDeckAppState extends State<FlutterDeckApp> {
             configuration: widget.configuration,
             router: _flutterDeckRouter,
             speakerInfo: widget.speakerInfo,
+            autoplayNotifier: _autoplayNotifier,
             controlsNotifier: _controlsNotifier,
             drawerNotifier: _drawerNotifier,
             markerNotifier: _markerNotifier,